Deleting member now functions [WARNING!]
This is a workaround, the oledb driver doesn't support deleting rows in
a linked file. So we updated the row with all NULL values, except the
ID.

The ID's are now hidden in the dataView, so are the empty rows.

This is just a nasty hack but it works... somewhat.